backend/requirements.txt
flask
flask-smorest
python-dotenv
pymongo
mongoengine
numpy
pandas
sentence-transformers==2.2.2
torch==2.1.0
flask_bcrypt
flask_jwt_extended
flask_cors
newsapi-python
boto3
requests==2.27.1